refactor: Clean up web app codebase
- Remove duplicate ThemePicker component (kept header version)
- Consolidate audio/playlist component structure
  - Remove old playlist component with broken SCSS reference
  - Flatten unnecessary audio component nesting
- Consolidate Sanity client configurations into single config.ts
  - Remove 4 redundant configuration files
  - Update all imports across 14+ files
- Improve overall code organization and maintainability
