+++ apps/web/src/components/rich-text/RichText.tsx
@@ -1,9 +1,9 @@
-import { config } from 'lib/sanityConfig';
+import { sanityConfig } from 'lib/sanity/config';
 import { PortableText } from '@portabletext/react';
 import { Components } from './Components';
 
 export const RichText = ({ value }) => {
 	return (
-		<PortableText value={value} components={Components} {...config} />
+		<PortableText value={value} components={Components} {...sanityConfig} />
 	)
 }
