--- apps/web/src/components/theme-picker/index.ts
@@ -1,1 +1,0 @@
-export { ThemePicker } from './theme-picker'
