--- apps/web/src/components/theme-picker/theme-picker.tsx
@@ -1,28 +1,0 @@
-'use client'
-
-import { useState, useEffect } from 'react'
-import { useTheme } from "next-themes"
-
-const ThemePicker = () => {
-  const [mounted, setMounted] = useState(false)
-  const { theme, setTheme } = useTheme()
-
-  useEffect(() => {
-    setMounted(true)
-  }, [])
-
-  if (!mounted) {
-    return null
-  }
-
-  return (
-    <div>
-      <select value={theme} onChange={e => setTheme(e.target.value)}>
-        <option value="dark">Dark</option>
-        <option value="light">Light</option>
-      </select>
-    </div>
-  )
-}
-
-export { ThemePicker }
