--- apps/web/src/lib/sanity/sanity.client.ts
@@ -1,19 +1,0 @@
-import { createClient } from 'next-sanity';
-
-export const projectId = process.env.NEXT_PUBLIC_SANITY_PROJECT_ID;
-export const dataset = process.env.NEXT_PUBLIC_SANITY_DATASET;
-export const apiVersion = process.env.NEXT_PUBLIC_SANITY_API_VERSION;
-
-export const sanityClient = createClient({
-	projectId: process.env.NEXT_PUBLIC_SANITY_PROJECT_ID,
-	dataset: process.env.NEXT_PUBLIC_SANITY_DATASET,
-	apiVersion: process.env.NEXT_PUBLIC_SANITY_API_VERSION,
-	useCdn: true,
-});
-
-export const sanityConfig = {
-	projectId: process.env.NEXT_PUBLIC_SANITY_PROJECT_ID,
-	dataset: process.env.NEXT_PUBLIC_SANITY_DATASET,
-	apiVersion: process.env.NEXT_PUBLIC_SANITY_API_VERSION,
-	useCdn: true,
-};
