--- apps/web/src/lib/sanity/sanityClient.ts
@@ -1,12 +1,0 @@
-import { createClient } from 'next-sanity';
-
-const projectId = process.env.NEXT_PUBLIC_SANITY_PROJECT_ID; // "pv8y60vp"
-const dataset = process.env.NEXT_PUBLIC_SANITY_DATASET; // "production"
-const apiVersion = process.env.NEXT_PUBLIC_SANITY_API_VERSION || '2023-05-03';
-
-export const sanityClient = createClient({
-	projectId,
-	dataset,
-	apiVersion, // https://www.sanity.io/docs/api-versioning
-	useCdn: true, // if you're using ISR or only static generation at build time then you can set this to `false` to guarantee no stale content
-});
