--- apps/web/src/lib/sanityConfig.ts
@@ -1,6 +1,0 @@
-export const config = {
-	dataset: process.env.NEXT_PUBLIC_SANITY_DATASET || 'production',
-	projectId: process.env.NEXT_PUBLIC_SANITY_PROJECT_ID,
-	apiVersion: '2021-10-21',
-	useCdn: process.env.NODE_ENV === 'production',
-};
